chore: renamed camelCase filenames to hyphen
